requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# OpenAI compatible client
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Data handling
datasets>=2.14.0
pyyaml>=6.0

# Optional: Progress tracking
tqdm>=4.65.0
